Upgraded to SIF Infrastructure 3.0.1
Upgraded the framework to use SIF Infrastructure 3.0.1. Updated the data
models to the latest version of the SIF AU 1.3 Data Model. Re-designed
the XML serialisation code to provide for better extensibility. Fixed an
issue whereby the root element of collections returned by Controllers
started with "ArrayOf". Made enhancements to the demo Setup.