Fixed urgent issue with demo AU Consumer
Fixed an issue introduced in version 0.17.0 whereby the demo AU Consumer
referenced a non-existant file - SifFramework.brokered.config. Updated
the demo AU Consumer and Provider to better reflect exception handling
and logging enhancements in the SifFramework library. Updated the demo
US Consumer and Provider to match the changes in the AU versions.