Added ability to connect to a SIF Broker
Updated the GenericConsumer to allow registration to a SIF Broker, as
well as (direct) to an Environment Provider. Updated the
GenericController to allow creation of Service Providers that can
connect to a SIF Broker, as well as run (directly) as an
Environment/Service Provider. Added functionality for Consumers and
Providers to store the session token (received after service
registration) locally so that state can be maintained between Consumer
and Provider sessions. Improved exception handling and logging in the
SifFramework library. Updated the demo Consumers and Providers to
reflect these changes.