Implement Query By Example
Updated the GenericConsumer to add a new Retrieve method that accepts an
"example" object. Updated the Get method of the GenericController to
handle a payload when a method override is requested. Updated the demo
AU Provider configuration to manage redirection when a method override
is requested. Updated the demo AU Consumer with an example call that
uses the new Retrieve method.